(vt) Initial renaming
